+++ src/Classes/CalcsTab.lua
@@ -1,4 +1,4 @@
--- Last Epoch Planner
+-- Last Epoch PlaBuildingnner
 --
 -- Module: Calcs Tab
 -- Calculations breakdown tab for the current build.