+++ src/Classes/CheckBoxControl.lua
@@ -1,4 +1,4 @@
--- Last Epoch Planner
+-- Last Epoch Building
 --
 -- Class: Check Box Control
 -- Basic check box control.