+++ src/Classes/FolderListControl.lua
@@ -1,4 +1,4 @@
--- Last Epoch Planner
+-- Last Epoch Building
 --
 -- Class: Folder List
 -- Folder list control.