+++ src/Classes/ModList.lua
@@ -1,4 +1,4 @@
--- Last Epoch Planner
+-- Last Epoch Building
 --
 -- Module: Mod List
 -- Stores modifiers in a flat list