+++ src/Classes/ModStore.lua
@@ -1,4 +1,4 @@
--- Last Epoch Planner
+-- Last Epoch Building
 --
 -- Module: Mod Store
 -- Base class for modifier storage classes