+++ src/Classes/SkillsTab.lua
@@ -1,4 +1,4 @@
--- Last Epoch Planner
+-- Last Epoch Building
 --
 -- Module: Skills Tab
 -- Skills tab for the current build.