+++ src/Modules/Build.lua
@@ -1,4 +1,4 @@
--- Last Epoch Planner
+-- Last Epoch Building
 --
 -- Module: Build
 -- Loads and manages the current build.